Check error codes not strings
